Change data import strategy
We can import data direcly as long as it does not need any node.js stuff. So anywhere we are using just the first level site[something], we can simply import a js object.

In places where we really need deeper data from yaml, we have to import through the server instead.