Switch to NPM, fix rollup, adjust config
Yarn was interminably stuck linking svelte-image to my local copy. Bye!

Once fixed, it was still clear that there are just some plugins that
will not play well with the svelte language server, and so I devised an
organizational structure that allows me to have one canonical config,
but where the offending packages can be added in a file that only goes
to the runtime version of Svelte Preprocess.

Other changes and additions are to support the tests that will appear in
the next commit.