Use DynamicCurrentSeasonImage to attempt to ensure dynamic images render correctly
I noticed today in production we were showing the wrong image for "opening today". Instead of Into the Woods, it was Lucky Stiff. The copy was all correct (which was an issue I previously addressed) but the image was wrong.

The reason is that the image is created at build time, but I don't know *exactly* when the image is set and unchangeable. Surely the change I have made to rendering logic via the new DynamicCurrentSeasonImage component will ensure that we can dynamically switch between all images.

What I don't know is what happens when the component itself is not rendered in the first pass on the server. I am fairly sure that the images will be created, but I don't know about the logic after that.

If each `<enhanced:img>` is somehow serialized by its parameters, then adding the DynamicCurrentSeasonImage anywhere on the page should allow me to dynamically add *another one* anywhere and it could use the same hash to resolve the rendered HTML for the image.

However, if the enhanced image is only referenced per instance of a component, then even that will not solve the problem of the image NOT appearing at build time, but then appearing later, at runtime.

So while this is strictly better... I don't know if it is either complete or if it is actually overkill.